BlockDates.jl

A Context aware fuzzy date matching tool

Package Features

Sum the squares of two numbers

Function Documentation

BlockDates.blockdate — Function

blockdate(
    txtin; 
    dtstart         = Date(0), 
    dtend           = Date(0),
    splits          = r"[\n\r]+",
    singleformats   = length(singleset),
    rangeformats    = length(rangeset),
    targetScore     = 5,
    verbose         = false)

    txtin         : Text block unput into blockdates
    dtstart       : Start date of dates
    dtend         : End date of dates  
    splits        : Defines the character used to split the input text. Default is line breaks    
    singleformats : Single formats allowed integers
    rangeformats  : Range formats allowed integers
    targetScore   : Cut score considered good enough
    verbose       : Provide extra feedback

BlockDates.dateformat2regex — Function

dateformat2regex converts input dateformats into potential regex matches.

BlockDates.replacemonth — Function

Replace month matches with numbers when found in text.

BlockDates.removedays — Function

Remove potential day name matches from text

BlockDates.removestopwords — Function

Remove potential stop words and common words from x: "NOTE|Between|On|At|Continued|Date" and "On|At|from"

BlockDates.removecommonprefix — Function

Remove common prefixes in the text.

BlockDates.removejoiner — Function

Remove joiner text: "thru|through|to"

BlockDates.singleGenerator — Function

singleGenerator(dateformat = "mm dd yyyy", scoremod = 5, mod = Day(0))

Generates a function that will interprets a single date from text passed to it.

Arguments

dateformat: date that the rangeInterpretor will attempt to match dateformat to look for are of the type used with the DateFormat() function

scoremod: score to assign to this range interpretation

mod: number of Days, Years to add to the date after interpretation

BlockDates.singleYearMissGen — Function

singleYearMissGen(infunction::Function, penalty = -3)

Missing Year in Date Single Allow for missing year in date eg. mm dd or dd mm

BlockDates.cleanDate — Function

cleanDate(text)

Make text more readable by the date interpreter. (make lowercase, replace month with #, remove days, remove common prefixes, remove stop words, remove joiners, replace markers used to divide dates, and remove common numeric ordering tags (st, nd, rd).

BlockDates.DateMod — Function

DateMod(inDate, scoremod; mods = "", defaultdate = Date(0))

Adjust dates to be within feasible range.

DateMod(inDate::Date, ender::Date, scoremod; mods = "", defaultdate = Date(0))

Apply DateMod to inDate and ender separately then do two more modifications.
If inDate is greater than ender swap.
If the date range is too wide (greater than 30 days) force the end date to be
only 6 days after the start date.

BlockDates.formatRange — Function

formatRange(textin; rangeformat = 1, defaultdate = Date(0))

Convert range from "11/29/2020 - 12/05/2020" to a dataframe with scoremods.

BlockDates.andSplit — Function

andSplit(txtin, splitters = "(?:\/)", ands = "(?:and|,)")

A function to split dates that have groupings like "and" and ",".

BlockDates.textToBlock — Function

textToBlock(txtsplit::Array{String}; 
                singleformat = 1, 
                rangeformat  = 1,
                defaultdate = Date(0))

Takes a date text string list and tries to match each element first as a range then as a single.

BlockDates.spreadOverNonDates — Function

spreadOverNonDates(inframe)

Merges rows with missing dates with previous row (unless first row then merges with next row).

BlockDates.joinDuplicateDates — Function

joinDuplicateDates(inframe, penalty = 7)

Join any duplicate dates which have different text together. Scoremod penalizes these observations by -7 number of duplicate dates.

BlockDates.fillMissing — Function

fillMissing(inframe, dtstart = Date(0), dtend   = Date(0))

Join any duplicate dates which have different text together. Scoremod penalizes these observations by -7 number of duplicate dates.

BlockDates.scoreDatesOutside! — Function

scoreDatesOutside!(inblock; dtstart, dtend, penalty = 5)

Scores dates outside of the dtstart - dtend range. Dates which are outside get penalized by the penalty value. If they are outside by more than 1 2 4 8 16 25 100 weeks then then accumulate another penalty for each step and the "outOfRange" gets a "*" appended to it for each step.

Parameters: inblock: DataFrame containing date, mod, and scoremod columns dtstart: the start of the interval of interest dtend: the end of the interval of interest penalty: the penalty to the scoremod for each step out of the dtstart/dtend range

BlockDates.dropBadMatches — Function

dropBadMatches(inframe::DataFrame, dropscore = -15)

Removes the date and score from any dates for which their interpretation is extremely low scored. This is most often helpful when a non-date becomes interpetted as a date.

Missing docstring.

Missing docstring for fixBlock. Check Documenter's build log for details.

BlockDates.scoreOutofOrder! — Function

scoreOutofOrder!(inblock, dtstart, dtend)

penalizes dates out of order

BlockDates.scoreBlock — Function

scoreBlock(inframe; dtstart = Date(0), dtend = Date(0))

Score the passed block by first adjusting for out of order entries then taking the average of the scoremod values.

using BlockDates, Dates

txt = """
03/14/2022 went for a walk
and a stroll
03/15 went sky diving
03/16/2022 went fishing
"""

blockdate(txt; dtstart = Date("2022-03-14"), dtend = Date("2022-03-16"))

3×8 DataFrame
 Row │ date        mods     scoremod  score    order    text                               singleformat  rangeformat 
     │ Date        String?  Int64?    Float64  Array…?  String                             Int64         Int64       
─────┼───────────────────────────────────────────────────────────────────────────────────────────────────────────────
   1 │ 2022-03-14                  2      2.0  [1]      03/14/2022 went for a walk \n an…             5            1
   2 │ 2022-03-15                  2      2.0  [3]      03/15 went sky diving                         5            1
   3 │ 2022-03-16                  2      2.0  [4]      03/16/2022 went fishing                       5            1